LayoutTests/svg/filters/feBlend-invalid-mode.xhtml

<html>
<body>
<p>PASS: Invalid blend modes do not trigger a crash.</p>
<svg xmlns="http://www.w3.org/2000/svg">
<defs>
    <filter id="filter">
        <feBlend id="blend" />
    </filter>
</defs>
<rect filter="url(#filter)" width="10" height="10"/>   
<script>
if (window.testRunner)
    testRunner.dumpAsText();
document.getElementById('blend').mode.baseVal=-1;
</script>
</svg>
</body>
</html>
